requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncssh==2.14.0
atpublic==4.0
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.1.0
blinker==1.6.3
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.6
dvc==3.27.0
dvc-data==2.18.2
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
ensure==1.0.2
entrypoints==0.4
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.10.0
funcy==2.0
gast==0.5.4
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grandalf==0.8
grpcio==1.59.0
gto==1.4.0
gunicorn==20.1.0
h5py==3.10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
isoduration==20.11.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.4.0
jupyter_server==2.9.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.14.0
kiwisolver==1.4.5
kombu==5.3.2
libclang==16.0.6
llvmlite==0.41.1
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mlflow==2.7.1
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2
notebook==7.0.6
notebook_shim==0.2.3
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
orjson==3.9.9
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pygit2==1.13.1
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-box==6.0.2
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0.1
pyzmq==25.1.1
querystring-parser==1.2.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
ruamel.yaml==0.18.2
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.3
scmrepo==1.4.0
seaborn==0.13.0
semver==3.0.2
Send2Trash==1.8.2
shap==0.43.0
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
slicer==0.0.7
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
sqltrie==0.8.0
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tomlkit==0.12.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
typer==0.9.0
types-python-dateutil==2.8.19.14
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
